refactor(symptom-presets): alias ALL_ABS_SYMPTOM_SUGGESTIONS to COMMON on web and mobile
